vp9: reset denoiser and cr-layer pars on resize
For spatial layers on config change in oxcf->width/height:
reset/re-alloc the denoiser and cr layer parameters.
This prevents crash for external resizing with SVC.

Also fix to setting of initial_width/height, needed to
support the feature in the issue below.

Added unittest for SVC with external resize.

Bug: b:392441185
Change-Id: Ibc2aaca1a786e7887c6527923e5abb654e59929c