vp9: Fixes to allow for external resize with SVC
For spatial layers on config change in oxcf->width/height:
reset/re-alloc the denoiser and cr layer parameters.
This prevents crash for external resizing with SVC.

Also fix to setting of initial_width/height, needed to
support the feature in the issue below, when the resize
is done on the very first frame.

Added unittest for SVC with external resize for (3SL, 3TL),
which triggers the crash without the fixes.
The test forces external resize several times during the
sequence, beginning with the very first frame, and sets
the layer bitrates to zero/non-zero to disable/enable
layers.

Bug: b:392441185
Change-Id: Ibc2aaca1a786e7887c6527923e5abb654e59929c