vp9: fix to tsan race for external resize
Set/update the stride values on a resized frame,
when new width is <= previous width.

Added new external resize tests, for single layer
and svc case, that trigger the same issue. Tests
run with 4 threads and small keyframe spacing. Issue
was occuring on keyframes after resolution went down
from 640x480 to 320x240, and is was triggered on 4x4
coding blocks in rd_pick_intra_mode_sb.

This issue was originally seen in a webrtc svc test
with the patch:
https://webrtc-review.googlesource.com/c/src/+/374661

Change-Id: I45211b773af808a317de0a4a1259ca17d695b8c6