vp8: Reset cpi->ref_framerate for very low framerate
cpi->ref_framerate can go down to zero or negative for
bad/inconsistent timestamp or duration inputs.
Unittest is attached as example. This can cause other
variables, like lc->avg_frame_size_for_layer, to overflow.

Fix is to reset cpi->ref_framerate in the function
vp8_new_framerate(), and move the call to
that function so temporal layers enters it.

The unittest  repros the issue in the bug attached.

Bug: 431520320
Change-Id: Id525bab2a01283c21c67eaff219064f1eee1df70