vp9/encoder: remove SOURCE_VAR_BASED_PARTITION
This partition search type has been unused since:
0ce51d823 experimental : partition using 1/8 x 1/8 image

This also removes the related `search_type_check_frequency` which is
set, but unused.

This fixes a clang-tidy warning: bugprone-sizeof-expression. If the code
is restored and needs to be fixed, the change required is
`sizeof(cpi->source_diff_var)` -> `sizeof(*cpi->source_diff_var)`.

Bug: 437631372
Change-Id: I7a185fefcabfae9fe399770bcfb9f0f98fb69b99