Fix to buffer overflow in vp9_pack_bitstream
The issue occurs on key frames in the nonrd intra
pickmode (sf->nonrd_keyframe = 1) for 4x4 blocks.
When the deadline is flipped from best/good_quality
to realtime_quality a keyframe is forced on that frame
with sf->nonrd_keyframe = 1 (this happens in the function
set_rt_speed_feature_framesize_independent()).

The nonrd intra pickmode does not correctly handle 4x4,
in particular the mi->bmi[] is not set. so it will take
the last value assigned, which can be an inter_mode if a
4x4 block was used on the previous delta frame. This then
causes intra_mode_encodings[] to take a non-intra mode
causing the buffer overflow. The conditions in the test
(in particular best_quality and  QP = 0) can trigger 4x4 on
delta frames.

Fix is to avoid 4x4 partitions on keyframes, unless
sf->nonrd_keyframe = 0.

Bug: b:441668134

Change-Id: I770c9937e6942da06a653a172b7aba061ec16e23